+++ src/class-settings.php
@@ -416,7 +416,7 @@ class Settings {
 				printf(
 					'<strong>%s</strong> <time datetime="%s">%s</time>',
 					esc_html__( 'Next run:', 'feed-consumer' ),
-					esc_attr( date_i18n( 'c', $next_run ) ),
+					esc_attr( wp_date( 'c', $next_run ) ),
 					esc_html(
 						sprintf(
 							/* translators: %s: Human readable time difference. */
@@ -442,8 +442,8 @@ class Settings {
 			printf(
 				'<p><strong>%s</strong> <time datetime="%s">%s</time></p>',
 				esc_html__( 'Last run:', 'feed-consumer' ),
-				esc_attr( date_i18n( 'c', $last_run ) ),
-				esc_html( date_i18n( get_option( 'date_format' ) . ' ' . get_option( 'time_format' ), $last_run ) ),
+				esc_attr( wp_date( 'c', $last_run ) ),
+				esc_html( wp_date( get_option( 'date_format' ) . ' ' . get_option( 'time_format' ), $last_run ) ),
 			);
 		}
 	}